AutoConf.pm: Fix false positive "compile_if_else" when CC=true
While typical AutoConf check_cc proves for compiler and whether compiler can
generate executables, this check is missing in ExtUtils::CBuilder. While
compile errors were detected by not receiving a file name for the generated
compilation, it seems that CBuilder doesn't prove the existence of that
result but deal with it as a flag.

Don't rely on assumption the compilation has been proved, prove.

Signed-off-by: Jens Rehsack <sno@netbsd.org>